requirements.txt
python-telegram-bot>=20.4
psycopg2-binary
apscheduler==3.10.4
python-dotenv